Skeleton project with spark and scala dependencies
The build.gradle specification is intentionally very thin. We are
interested in seeing a hand-written implementation for discussion, but
if bringing in an external library allows for something that would
be useful in solving the problem but otherwise take longer than a few
hours to implement manually, go ahead and add it with documentation
about what is being brought in and why.